+++ frontend/src/util/middleware/socketActions/modal.tsx
@@ -3,6 +3,7 @@
  */
 import { store } from "../../../app/store";
 import { setModal } from "../../../app/features/modalSlice";
+import { User } from "../../../app/features/types";
 
 export const setModalInviteToChat = (peerId: string, peerUsername: string) => {
   const modalData = {
@@ -15,5 +16,18 @@ export const setModalInviteToChat = (peerId: string, peerUsername: string) => {
     socketEvent: 'invite private chat'
   }
 
+  store.dispatch(setModal(modalData))
+}
+
+export const setModalInviteRequest = (inviterId: string, inviter: User) => {
+  const modalData = {
+    modalName: 'private chat request',
+    modalContent: `${inviter.username} has invited you to a private chat?`,
+    confirmBtnText: 'Yes, accept invite.',
+    declineBtnText: 'No, decline invite.',
+    isActive: true,
+    peerId: inviterId,
+    socketEvent: 'invite requested'
+  }
   store.dispatch(setModal(modalData))
 }
